refactor(tab-manager): move orphaned command palette items out of empty directory
