+++ apps/tab-manager/src/lib/components/command-palette-items.ts
